# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: La Escoba Card Game API
  description: Microservices backend for the Spanish card game "La Escoba".
  version: 1.0.0
  contact:
    name: ASE Project Team
servers:
  - url: https://localhost:5000
    description: Local Development Server (HTTPS Gateway)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /players/{username}:
    post:
      summary: Initialize player profile
      tags: [Player]
      parameters:
        - in: path
          name: username
          schema: { type: string }
          required: true
      responses:
        '201':
          description: Profile created
    get:
      summary: Get player statistics
      tags: [Player]
      parameters:
        - in: path
          name: username
          schema: { type: string }
          required: true
      responses:
        '200':
          description: Player stats returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  username: { type: string }
                  matches_played: { type: integer }
                  matches_won: { type: integer }
                  total_score: { type: integer }
                  win_rate: { type: number }

  # --- CARDS SERVICE ---
...